requirements.txt
flask
pdf2docx
docx2pdf
python-docx
pymupdf
pdfplumber